test: yarn lock header parsing test
